almost done with bktrees, added some more tests for strings, need to find a good edit distance tolerance

+++ src/bkTrees/appMatching/editDistance.h
@@ -3,6 +3,6 @@
 
 #include "../../string.h"
 
-int getEdit(simple::string f, simple::string s);
+unsigned long getEdit(simple::string f, simple::string s);
 
 #endif // !editDistance
